+++ CHANGELOG.md
@@ -5,7 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [2.4.0] - ?
 ### Added
  - MantaSeekableByteChannel now extends from InputStream.
-
+ - Added move method to MantaClient. 
+ 
 ## [2.3.0] - 2016-04-22
 ### Changed
  - Upgraded http-signatures to 2.2.0 so that we are keeping pace with upstream